requirements.txt
solax @ git+https://github.com/GustavoIO/solax-add-x1-hybrid-5@master
paho-mqtt
